requirements.txt
# Core dependencies for VRM Framework
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0
tensorboard>=2.13.0
accelerate>=0.20.0
